+++ packages/core/src/types/IStreamingVolumeProperties.ts
@@ -1,10 +1,6 @@
-import { VolumeTimePoints } from './IImageVolume';
-
 interface IStreamingVolumeProperties {
   /** imageIds of the volume  */
-  imageIds?: Array<string>;
-  /** time points data for 4D volumes  */
-  timePointsData?: VolumeTimePoints;
+  imageIds: Array<string>;
 
   /** loading status object for the volume containing loaded/loading statuses */
   loadStatus: {
