+++ packages/core/src/types/index.ts
@@ -3,10 +3,10 @@
 import type ICamera from './ICamera';
 import type IEnabledElement from './IEnabledElement';
 import type ICache from './ICache';
-import type IVolume from './IVolume';
+import type { IVolume, VolumeScalarData } from './IVolume';
 import type { VOI, VOIRange } from './voi';
 import type ImageLoaderFn from './ImageLoaderFn';
-import type { IImageVolume, VolumeTimePoints } from './IImageVolume';
+import type IImageVolume from './IImageVolume';
 import type VolumeLoaderFn from './VolumeLoaderFn';
 import type IRegisterImageLoader from './IRegisterImageLoader';
 import type IStreamingVolumeProperties from './IStreamingVolumeProperties';
@@ -80,9 +80,9 @@ export type {
   IEnabledElement,
   ICache,
   IVolume,
+  VolumeScalarData,
   IViewportId,
   IImageVolume,
-  VolumeTimePoints,
   IRenderingEngine,
   ScalingParameters,
   PTScaling,
